requirements.txt
# 🌐 Core RAG Dependencies
langchain
langchain-community
langchain-huggingface
chromadb
openai
tiktoken
PyMuPDF
python-dotenv
sentence-transformers

# 💬 Together.ai LLM Integration
together

# 🖥️ Streamlit App UI
streamlit

# 🧪 Optional: Notebook + Dev Tools
ipykernel
jupyter
